chore: rename webnames to webnamesru
