+++ providers/dns/webnamesru/internal/types.go
